Add a CONTRIBUTING.md
This is based on wiki pages and DP's CONTRIBUTING.md and it adds some
policies that previously were expected but not documented.

This is intended to replace policy and licensing sections of the wiki,
partly to collate these and partly because only the Team should be able
to push updates to them.

Signed-off-by: bones_was_here <bones_was_here@xonotic.au>