Windows: use build-environment-supplied libSDL2 in developer builds
This requires developers to have installed SDL2 from MSYS2's `pacman` as
documented in https://gitlab.com/xonotic/xonotic/-/wikis/Repository_Access

Fixes failure to build current DP on Windows: the precompiled SDL2 in
this repo is too old.
Removes the need for manually maintaining that library and storing it in
this repo.