rsync: improve status messages, fix target selection
Fixes printing the wrong message "release" when updating from an
autobuild zip. That code can't distinguish between rsync and release
reliably (because an autobuild zip has the release-style pk3 names) and
doesn't need to.

Fixes the "secret trick" potentially clobbering the filesystem:
downloading update-to-autobuild.sh separately to /home/user/Downloads/
and running from there it downloaded Xonotic to ~/Downloads/Xonotic/,
but if run again it detected that new Xonotic/ directory and set the
rsync target to ../../.. (ie /). That fallback target is no longer
supported as it's a bit odd to download a complete copy via the updater,
but it can still be done by creating the script parent directories.

Moves the main "Do you want to continue" after other messages so the
user is better informed about what they're confirming.