Makefile: support downstream packages
This is written with reference to the old Makefile (but with updates and
fixes) and some 0.8.6 distro packages. It aims to support the way they want
to package Xonotic to reduce the amount of fragile version-specific
scripting they do so they have less problems building new versions.

- Supports several GNU make "standard" directory variables and targets
- Supports omitting the xonotic-linux-*.sh files and symlinks by setting
  BASEDIR
- Supports client-only and server-only installs
- Supports renaming desktop and icon files (for flathub)
- Installs data to $(prefix)/share/xonotic instead of
  $(prefix)/lib/xonotic to match GNU make docs and current distro
  packages.
- Includes metadata and CA key for player IDs

Some documentation is added at https://gitlab.com/xonotic/xonotic/-/wikis/Compiling